Move smoke test into battleStateProgress test
